+++ manual/server-examples/src/docs/dev/plugins-extensions-desktop.asciidoc
@@ -1,12 +1,8 @@
 [[plugins-extensions-desktop]]
-Installing Plugins and Extensions in Neo4j Desktop
-==================================================
+= Installing Plugins and Extensions in Neo4j Desktop
 
-Neo4j Desktop can also be extended with server plugins and extensions. Neo4j
-Desktop will add all jars in '%ProgramFiles%\Neo4j Community\plugins' to the
-classpath, but please note that nested directories for plugins are currently
-not supported.
+Neo4j Desktop can also be extended with server plugins and extensions.
+Neo4j Desktop will add all jars in '%ProgramFiles%\Neo4j Community\plugins' to the classpath, but please note that nested directories for plugins are currently not supported.
 
 Otherwise server plugins and extensions are subject to the same rules as usual.
 Please note when configuring server extensions that _neo4j.conf_ for Neo4j Desktop lives in '%APPDATA%\Neo4j Community'.
-
