+++ manual/server-examples/src/docs/dev/server-unmanaged-extensions.asciidoc
@@ -1,40 +1,42 @@
 [[server-unmanaged-extensions]]
-Unmanaged Extensions 
-====================
+= Unmanaged Server Extensions
 
 Sometimes you'll want finer grained control over your application's interactions with Neo4j than cypher provides.
 For these situations you can use the unmanaged extension API.
 
 [CAUTION]
+--
 This is a sharp tool, allowing users to deploy arbitrary http://en.wikipedia.org/wiki/JAX-RS[JAX-RS] classes to the server so be careful when using this.
 In particular it's easy to consume lots of heap space on the server and degrade performance.
 If in doubt please ask for help via one of the community channels (see <<preface>>).
+--
 
 :leveloffset: 3
+
 [[intro-unmanaged-extensions]]
 = Introduction to unmanaged extensions
 The first step when writing an unmanaged extension is to create a project which includes dependencies to the JAX-RS and Neo4j core jars.
 In Maven this would be achieved by adding the following lines to the pom file:
 
-[source,xml]
---------
+[source, xml]
+----
 <dependency>
     <groupId>javax.ws.rs</groupId>
     <artifactId>javax.ws.rs-api</artifactId>
     <version>2.0</version>
     <scope>provided</scope>
 </dependency>
---------
+----
 
-["source","xml","unnumbered","2",presubs="attributes"]
---------
+["source", "xml", "unnumbered", "2", presubs="attributes"]
+----
 <dependency>
     <groupId>org.neo4j</groupId>
     <artifactId>neo4j</artifactId>
     <version>{neo4j-version}</version>
     <scope>provided</scope>
 </dependency>
---------
+----
 
 Now we're ready to write our extension.
 
@@ -42,7 +44,7 @@ In our code we'll interact with the database using +GraphDatabaseService+ which 
 The following examples serves as a template which you can base your extension on:
 
 .Unmanaged extension example
-[snippet,java]
+[snippet, java]
 ----
 component=neo4j-server-examples
 source=org/neo4j/examples/server/unmanaged/HelloWorldResource.java
@@ -64,10 +66,10 @@ org.neo4j.server.thirdparty_jaxrs_classes=org.neo4j.examples.server.unmanaged=/e
 
 Our hello method will now respond to +GET+ requests at the URI: +http://\{neo4j_server}:\{neo4j_port}/examples/unmanaged/helloworld/\{nodeId}+. e.g.
 
-[source,bash]
------
+[source, bash]
+----
 curl http://localhost:7474/examples/unmanaged/helloworld/123
------
+----
 
 which results in
   
@@ -75,11 +77,6 @@ which results in
 ----
 Hello World, nodeId=123
 ----
-
-[CAUTION]
-Whilst Neo4j is tested to run on JVM 8, Neo4j server is currently compiled using
-JDK 7, to ensure compatibility for JVM 7 deployments. When compiling plugins for
-Neo4j Server, we strongly recommend using JDK 7 also.
 
 :leveloffset: 3
 
@@ -89,7 +86,7 @@ include::server-unmanaged-extensions-streaming.asciidoc[]
 
 include::server-unmanaged-extensions-cypher-execution.asciidoc[]
 
-:leveloffset: 2
+:leveloffset: 3
 
 include::server-testing.asciidoc[]
